rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastembed>=0.7.3",
    "langchain-google-genai>=3.1.0",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "ruff>=0.14.5",
    "typhoon-ocr>=0.4.1",
]
